requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3
botocore
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.10
django-allauth==0.61.1
django-bootstrap-v5==1.0.11
django-debug-toolbar==4.2.0
django-on-heroku==1.1.2
django-storages==1.14.2
ffmpeg-python==0.2.0
future==0.18.3
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.2
procfile==0.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3
whitenoise==6.6.0
django-heroku
django-notifications-hq==1.8.3
setuptools
setuptools